break down roleTable component
