+++ src/components/roleAssignment/roleTable/TableContent.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { userDataType } from "../../../contexts/AuthContext";
+import { getRole } from "../../../utils/Helpers";
+
+interface TableContentProps {
+  enteries: userDataType[];
+}
+
+const TableContent = ({ enteries }: TableContentProps) => {
+  console.log(enteries);
+
+  return (
+    <tbody>
+      {enteries.map((userData: userDataType, index: number) => (
+        <tr key={index}>
+          <td data-cell="name">{userData.displayName}</td>
+          <td data-cell="email">{userData.email}</td>
+          <td data-cell="role">{getRole(userData)}</td>
+          <td data-cell="date">{userData.createdAt}</td>
+        </tr>
+      ))}
+    </tbody>
+  );
+};
+
+export default TableContent;
