--- src/utils/FormatDate.ts
@@ -1,9 +1,0 @@
-export const formatDate = (date: Date) => {
-  return date
-    .toLocaleDateString("en-us", {
-      year: "numeric",
-      month: "numeric",
-      day: "numeric",
-    })
-    .toString();
-};
